+++ androidApp/src/main/java/com/example/tasky/android/login/di/RegisterModule.kt
@@ -1,8 +1,8 @@
 package com.example.tasky.android.login.di
 
 import com.example.tasky.android.login.viewmodel.RegisterViewModel
+import com.example.tasky.login.data.LoginRepository
 import com.example.tasky.login.domain.ILoginRepository
-import com.example.tasky.login.domain.LoginRepository
 import org.koin.core.module.dsl.singleOf
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.bind
